// src/app/models/personal.model.ts
export class Personal{
    emailId: string;
    insuranceId: string;
    firstName: string;
    middleName: string;
    lastName: string;
    dob: Date;
    mobileNumber: string;
    gender: string;
    address: string;
    city: string;
    state: string;
    county: string;
    zipCode: string;
    alternateNumber: string
}